requirements.txt
groq
pydantic
requests
python-dotenv
PyGithub
slack_sdk
sendgrid
streamlit
rich
openai